Fixes for post-install script

+++ post-install.sh
@@ -1,9 +1,11 @@
-#!/bin/bash 
+#!/bin/bash
 
 set -e
 
 # Fix terminal for Tilix
-sudo ln -s /etc/profile.d/vte-2.91.sh /etc/profile.d/vte.sh
+if [ ! -e /etc/profile.d/vte.sh ]; then
+	sudo ln -s /etc/profile.d/vte-2.91.sh /etc/profile.d/vte.sh
+fi
 
 # Add KeepassXC PPA
 sudo add-apt-repository --yes --no-update --enable-source ppa:phoerious/keepassxc
@@ -12,6 +14,7 @@ sudo add-apt-repository --yes --no-update --enable-source ppa:phoerious/keepassx
 sudo add-apt-repository --yes --no-update --enable-source ppa:cdemu/ppa
 
 # Add Docker repository
+sudo apt install --yes curl
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
 sudo add-apt-repository --yes --no-update --enable-source "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
 
@@ -29,19 +32,20 @@ wget -q https://www.virtualbox.org/download/oracle_vbox.asc -O- | sudo apt-key a
 sudo apt-add-repository --yes --no-update --enable-source "deb https://download.virtualbox.org/virtualbox/debian bionic contrib"
 
 # Add Lutris repository
-ver=$(lsb_release -sr); if [ $ver != "18.04" -a $ver != "16.04" ]; then ver=18.04; fi 
+ver=$(lsb_release -sr); if [ $ver != "18.04" -a $ver != "16.04" ]; then ver=18.04; fi
 echo "deb http://download.opensuse.org/repositories/home:/strycore/xUbuntu_$ver/ ./" | sudo tee /etc/apt/sources.list.d/lutris.list
 wget -q http://download.opensuse.org/repositories/home:/strycore/xUbuntu_$ver/Release.key -O- | sudo apt-key add -
 
 # Install packages
-sudo apt update
-sudo apt full-upgrade
-sudo apt install git keepassxc apt-file pavucontrol htop gimp chrome-gnome-shell vim-gnome synaptic \
+sudo apt --yes update
+sudo apt --yes full-upgrade
+sudo apt --yes install git keepassxc apt-file pavucontrol htop gimp chrome-gnome-shell vim-gnome synaptic \
     lm-sensors kdeconnect gcdemu python-dev python3-dev python3-pip silversearcher-ag lxd lxd-tools \
-    nmap ncdu docker-ce ctags postgresql gnash libvulkan-dev libpng-dev libwayland-dev libxcb1-dev \
-    meson ninja w3m devscripts git-buildpackage libgirepository1.0-dev debhelper mono-complete \
+    nmap ncdu docker-ce exuberant-ctags postgresql gnash libvulkan-dev libpng-dev libwayland-dev libxcb1-dev \
+    meson ninja-build w3m devscripts git-buildpackage libgirepository1.0-dev debhelper mono-complete \
     neofetch jstest-gtk sc-controller mesa-utils mesa-vulkan-drivers linux-tools-common linux-tools-generic \
-    winehq-staging gimp inkscape lutris virtualbox-5.2 tilix
+    winehq-staging gimp inkscape lutris virtualbox-5.2 tilix zsh cmake \
+    gconf-service gconf-service-backend gconf2-common libappindicator1 libc++1 libc++abi1 libgconf-2-4 libindicator7
 
 # Install Chrome
 wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
@@ -49,7 +53,7 @@ sudo dpkg -i google-chrome-stable_current_amd64.deb
 rm google-chrome-stable_current_amd64.deb
 
 # Install Discord
-https://dl.discordapp.net/apps/linux/0.0.5/discord-0.0.5.deb
+wget https://dl.discordapp.net/apps/linux/0.0.5/discord-0.0.5.deb
 sudo dpkg -i discord-0.0.5.deb
 rm discord-0.0.5.deb
